Address PR comment about not useful inline comment.
